@vraman3: Saving Joint Data
LINE 182: At the start of MainWindow() I get the current System time ->
"startTime".

Since we are tracking the body position and then sending the OSC signals
based on that, it made sense to me to write the joint data at the same
time.

We need to pass that "startTime" to SendInstrumentData().
So we need to pass from MainWindow() to Update, to SendInstrumentData()
etc etc.

I can't test it to see how to iterate through those joints. So I just
did the lame way and put in 20 statements, one for each joint.

The writing stuff starts at LINE 498