Added joint data, removed FBX (no longer using)
Added a short capture data clip in CSV format. It represents a person
waving their arms, jumping up and down, and then kicking the right leg,
in the timespan of roughly 4 seconds.

These are the column values:
JointName,JointXPos,JointYPos,JointZPos,elapsedTimeInMilliseconds

The timestamp is fairly arbitrary, but just subtract the first time to
set everything relative to 0 and you should have a good animation.

Note that position values are in meters.